+++ packages/vsce/__tests__/__unit__/doc/meta.localfile.test.ts
@@ -35,7 +35,6 @@ describe("Local File Meta", () => {
       update: "UPDATABLE",
       add: "ADDABLE",
       delete: "DELETABLE",
-      recovstatus: "RECOVERABLE"
     });
   });