+++ packages/vsce/__tests__/__unit__/doc/meta.localfile.test.ts
@@ -32,6 +32,9 @@ describe("Local File Meta", () => {
       recordsize: "100",
       status: "ENABLED",
       vsamtype: "VSAM",
+      update: "UPDATABLE",
+      add: "ADDABLE",
+      delete: "DELETABLE",
     });
   });
 
@@ -145,7 +148,8 @@ describe("Local File Meta", () => {
       },
       {
         key: "Permission",
-        value: `READABLE, BROWSABLE`,
+        value: `BROWSABLE, READABLE, UPDATABLE, 
+                ADDABLE, DELETABLE`,
       },
       {
         key: "Key length",